refactor: move ValidationRules to Concerns
